requirements.txt
click==6.2
tabulate==0.7.5
GitPython==1.0.1
